requirements_dev.txt
#pgcli
-r requirements.txt
black>=19.10b0
codecov
coverage
django-debug-toolbar~=2.0
django_factory_boy
factory_boy
flake8
flake8-assertive
flake8-bugbear
flake8-comprehensions
flake8-django
ipython-sql
pdbpp>=0.9.2
pluggy>=0.7
pyinstrument
pytest-cov>=2.5.
pytest-django>=3.1.2
pytest-mock>=1.10.0
pytest-sugar>=0.9.1
pytest-xdist
pytest>=3.6.2
termcolor==1.1.0
wdb
